correct type for getRecursorThreadId
(cherry picked from commit 30da2030cf0cc0c5f9d59b6e30560c381a847432)

+++ pdns/pdns_recursor.cc
@@ -244,9 +244,9 @@ ArgvMap &arg()
   return theArg;
 }
 
-unsigned int getRecursorThreadId()
+int getRecursorThreadId()
 {
-  return static_cast<unsigned int>(t_id);
+  return t_id;
 }
 
 int getMTaskerTID()
